1: extract the creation of the `app` into a newly-created `./app.js`, which is a prerequisite for being able to write tests with Jest and SuperTest
